app/container/requirements.txt
omegaconf==2.3.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
torch==2.7.1
whisperx==3.4.2
